chore: alarmTime 자료형 변경

+++ backend/src/main/java/com/pppppp/amadda/schedule/entity/Participation.java
@@ -37,5 +37,5 @@ public class Participation extends BaseEntity {
 	private String scheduleMemo;
 
 	@Column
-	private Long alarmTime;
+	private Integer alarmTime;
 }
